+++ package.json
@@ -14,7 +14,6 @@
     "dotenv": "^6.1.0",
     "express": "^4.16.3",
     "hubot-rpc-gen": "^0.1.1",
-    "request": "^2.88.0",
     "unique-random": "^2.0.0"
   },
   "devDependencies": {
